+++ src/main/java/net/frozenblock/trailiertales/tag/TTBiomeTags.java
@@ -13,6 +13,7 @@ public class TTBiomeTags {
 	public static final TagKey<Biome> HAS_JUNGLE_RUINS = bind("has_structure/jungle_ruins");
 	public static final TagKey<Biome> HAS_SAVANNA_RUINS = bind("has_structure/savanna_ruins");
 	public static final TagKey<Biome> HAS_DEEPSLATE_RUINS = bind("has_structure/deepslate_ruins");
+	public static final TagKey<Biome> HAS_SNOWY_RUINS = bind("has_structure/snowy_ruins");
 	public static final TagKey<Biome> HAS_RUINS = bind("has_structure/ruins");
 
 	@NotNull