+++ CHANGELOG.md
@@ -5,6 +5,8 @@ hi
 -----------------
 - Added 42 new rooms for Catacombs.
 - Chains now generate more commonly in Catacombs.
+- Added 14 new tall pieces for Generic Ruins.
+- Added new ruins for snowy biomes.
 - Fixed a vanilla bug where Powered Rails, Activator Rails, and Detector Rails wouldn't rotate properly in structure generation.
 - Cracked and Chiseled Purpur Blocks can now be broken faster with a pickaxe. ([#10](https://github.com/FrozenBlock/TrailierTales/issues/10))
 - Fixed a crash where modded boats that aren't implemented the same way as vanilla cause a crash. ([#9](https://github.com/FrozenBlock/TrailierTales/issues/9))